@@ -1,7 +1,7 @@
 {
   "name": "bigspring-light-nextjs",
   "description": "a nextjs and tailwindcss SAAS starter template",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "private": true,
   "license": "MIT",
   "scripts": {
@@ -15,13 +15,14 @@
     "github-slugger": "^2.0.0",
     "gray-matter": "^4.0.3",
     "marked": "^4.2.4",
-    "next": "13.0.6",
+    "next": "^14.2.7",
     "next-mdx-remote": "^4.2.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-gtm-module": "^2.0.11",
     "react-icons": "^4.7.1",
     "react-lite-youtube-embed": "^2.3.52",
+    "rehype-highlight": "^7.0.0",
     "rehype-slug": "^5.1.0",
     "remark-gfm": "^3.0.1",
     "swiper": "^8.4.5"
