@@ -1,23 +1,34 @@
 import Pagination from "@components/Pagination";
 import config from "@config/config.json";
-import Base from "@layouts/Baseof";
+import SeoMeta from "@layouts/SeoMeta";
 import { getListPage, getSinglePage } from "@lib/contentParser";
-import { parseMDX } from "@lib/utils/mdxParser";
 import { markdownify } from "@lib/utils/textConverter";
 import Posts from "@partials/Posts";
 const { blog_folder } = config.settings;
 
 // blog pagination
-const BlogPagination = ({ postIndex, posts, currentPage, pagination }) => {
+const BlogPagination = async ({ params }) => {
+  //
+  console.log("🪲 :", params);
+  const currentPage = parseInt((params && params.slug) || 1);
+  const { pagination } = config.settings;
+  const posts = await getSinglePage(`content/${blog_folder}`).sort(
+    (post1, post2) =>
+      new Date(post2.frontmatter.date) - new Date(post1.frontmatter.date)
+  );
+  const postIndex = await getListPage(`content/${blog_folder}/_index.md`);
+
+  //
   const indexOfLastPost = currentPage * pagination;
   const indexOfFirstPost = indexOfLastPost - pagination;
   const totalPages = Math.ceil(posts.length / pagination);
   const currentPosts = posts.slice(indexOfFirstPost, indexOfLastPost);
-  const { frontmatter, content } = postIndex;
+  const { frontmatter } = postIndex;
   const { title } = frontmatter;
 
   return (
-    <Base title={title}>
+    <>
+      <SeoMeta title={title} />
       <section className="section">
         <div className="container">
           {markdownify(title, "h1", "h1 text-center font-normal text-[56px]")}
@@ -29,15 +40,14 @@ const BlogPagination = ({ postIndex, posts, currentPage, pagination }) => {
           />
         </div>
       </section>
-    </Base>
+    </>
   );
 };
 
 export default BlogPagination;
 
-// get blog pagination slug
-export const getStaticPaths = () => {
-  const getAllSlug = getSinglePage(`content/${blog_folder}`);
+export async function generateStaticParams() {
+  const getAllSlug = await getSinglePage(`content/${blog_folder}`);
   const allSlug = getAllSlug.map((item) => item.slug);
   const { pagination } = config.settings;
   const totalPages = Math.ceil(allSlug.length / pagination);
@@ -45,36 +55,11 @@ export const getStaticPaths = () => {
 
   for (let i = 1; i < totalPages; i++) {
     paths.push({
-      params: {
-        slug: (i + 1).toString(),
-      },
+      slug: (i + 1).toString(),
     });
   }
 
-  return {
-    paths,
-    fallback: false,
-  };
-};
+  console.log("🪲 :", paths);
 
-// get blog pagination content
-export const getStaticProps = async ({ params }) => {
-  const currentPage = parseInt((params && params.slug) || 1);
-  const { pagination } = config.settings;
-  const posts = getSinglePage(`content/${blog_folder}`).sort(
-    (post1, post2) =>
-      new Date(post2.frontmatter.date) - new Date(post1.frontmatter.date)
-  );
-  const postIndex = await getListPage(`content/${blog_folder}/_index.md`);
-  const mdxContent = await parseMDX(postIndex.content);
-
-  return {
-    props: {
-      pagination: pagination,
-      posts: posts,
-      currentPage: currentPage,
-      postIndex: postIndex,
-      mdxContent: mdxContent,
-    },
-  };
-};
+  return paths;
+}
